pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bilibili_downloader"
version = "0.1.0"
description = "哔哩哔哩视频下载、转写和摘要生成工具"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "requests>=2.28.0",
    "pydantic>=2.0.0",
    "oss2>=2.17.0",
    "pycryptodome>=3.17.0",
    "python-dotenv>=1.0.0",
    "aliyun-python-sdk-core>=2.13.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-mock>=3.10.0",
    "ruff>=0.0.278",
    "mypy>=1.0.0",
]

[tool.ruff]
select = ["E", "F", "W", "I"]
line-length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py" 